requirements.txt
torch
marker-pdf
ollama
python-dotenv
